Set default maxLandingPadSize when we know station type
This should help with it sometimes being null (when we know it's large).

I wasn't 100% all Surface Outposts were large (as Planetary Ports are) but I checked the data and none of them historically (out of tens of thousands) are not large to seems a safe bet in the absence of explicit info otherwise.