Update dependancies
It seems like there is now a non-beta version of the zeromq package that is compatible with the EDDN output (or maybe the EDDN service has been updated, I'm not sure).

Either way tested this and it seems to work without having to use an older beta release now.

I've also explicitly specific the version of Node.js used in production (and consequently have tested against).