Improvements to maintenance routine
Now only backs up the systems DB (which is 25 GB) on a weekly basis instead of daily.

This should reduce the daily maintenance window to about 5 mintes (during which time the site still works but new data is not ingested).

I might move all the major maintenance and backup publication  to weekly and algin it with the game downtime.