Refacctor timing for backups and stat generation
This way the tasks will not collide and will stop hitting the DB at the same time,, and the backups should be online sooner.

The commodity stats are an entirely backend task used for things that don't change radically often, like average pricing and commodity supply/demand, so can be lower priority.