Tweak to default pragmas for databases
* Changing sync mode from FULL (default) to NORMAL
* Setting busy timeout to 5 seconds

The flrst slightly improves write performanc (at the cost of dataloss in the event of a crash).

The seconds provides more robustness in the even of a maintenance script or other query I might have to run manually takes a bit longer to complete.

Both are required options for Litestream replication which I'm consdering exploring, I would first like to understand if there would be any issues with these options in practice.